Fix handler mode behavior
The previous implementation made no sense. In general, if an application
project defines a handler, there is no way a library it depends on could
know that definition (that would invert the dependency relationship).
Therefore this commit defines one known violation handler entry point
to be implemented by an application. This also means that handlers do
not (and can not) support constexpr evaluation. Instead, it now falls
back to assert() in constant evaluations.