Turn exceptions during contract evaluation into contract failures
Previously, if the contract condition caused an exception to be thrown,
that exception would escape into the parent scope, and from there
possibly to the caller. This meant that the contract checking mode could
add new failure cases (beyond the ones configured via the mode).
Now, exceptions thrown during contract evaluation result in a contract
failure, and the exception's what() message might be used as an
additional diagnostic.